dokoncit nacitani ikon

+++ src/Controller/Admin/UploadController.php
@@ -2,9 +2,6 @@
 
 namespace App\Controller\Admin;
 
-
-
-use App\Entity\Foto;
 use App\Entity\FotoInterface;
 use Doctrine\ORM\EntityManagerInterface;
 use Intervention\Image\Drivers\Imagick\Driver;
@@ -14,13 +11,12 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Attribute\Route;
 
-class GalerieUploadController extends AbstractController
+class UploadController extends AbstractController
 {
-    public function __construct(private readonly array $mapEntity, private readonly array $mapSetter, private readonly array $mapFileEntity, private readonly array $mapUploadDir)
+    public function __construct(private readonly array $mapEntity, private readonly array $mapSetter, private readonly array $mapFileEntity, private readonly array $mapUploadDir, private readonly array $mapGetter)
     {
 
     }
-
 
     #[Route('/admin/upload/{data_type}/upload/{entity}/{id}', name: 'admin_upload_upload', methods: ['POST'])]
     public function upload(
@@ -30,22 +26,23 @@ class GalerieUploadController extends AbstractController
         Request $request,
         EntityManagerInterface $em
     ): JsonResponse {
+        $config = [
+            'fileEntity' => $this->mapFileEntity[$data_type] ?? null,
+            'uploadDir'  => $this->mapUploadDir[$data_type] ?? 'files',
+            'class' => $this->mapEntity[$entity] ?? null,
+            'setter' => $this->mapSetter[$entity] ?? null,
+        ];
         $file = $request->files->get('file');
-
         if (!$file) {
             return $this->json(['error' => 'No file'], 400);
         }
 
-        $class = $this->mapEntity[$entity] ?? null;
-        $setter = $this->mapSetter[$entity] ?? null;
-        $fileEntity = $this->mapFileEntity[$data_type] ?? null;
-        $uploadDir = $this->mapUploadDir[$data_type] ?? 'files';
 
-        if (!$class || !$setter || !$fileEntity) {
+        if (!$config['class'] || !$config['setter'] || !$config['fileEntity']) {
             return $this->json(['error' => 'Unknown entity'], 400);
         }
 
-        $object = $em->getRepository($class)->find($id);
+        $object = $em->getRepository($config['class'])->find($id);
 
 
         if (!$object) {
@@ -53,7 +50,7 @@ class GalerieUploadController extends AbstractController
         }
 
         // ===== FILE SYSTEM =====
-        $baseDir = $this->getParameter('kernel.project_dir') . '/public/uploads/'.$uploadDir;
+        $baseDir = $this->getParameter('kernel.project_dir') . '/public/uploads/'.$config['uploadDir'];
 
         if (!is_dir($baseDir)) {
             mkdir($baseDir, 0777, true);
@@ -86,14 +83,14 @@ class GalerieUploadController extends AbstractController
         }
 
         // uložit do DB
-        $new_file = new $fileEntity;
-        $new_file->setSoubor('uploads/'. $uploadDir.'/' . $filename);
+        $new_file = new $config['fileEntity'];
+        $new_file->setSoubor('uploads/'. $config['uploadDir'].'/' . $filename);
         //$foto->setNazev($file->getClientOriginalName());
         $new_file->setNazev('');
         $new_file->setPosition(0);
 
         // dynamické přiřazení
-        $new_file->{$setter}($object);
+        $new_file->{$config['setter']}($object);
 
         $em->persist($new_file);
         $em->flush();
@@ -101,12 +98,15 @@ class GalerieUploadController extends AbstractController
         $response = [
             'id' => $new_file->getId(),
             'name' => '(' . $file->getClientOriginalName() . ')',
-            'url' => 'uploads/'. $uploadDir.'/' . $filename,
+            'url' => 'uploads/'. $config['uploadDir'].'/' . $filename,
         ];
 
         if ($data_type == 'image') {
-            $response['thumbUrl'] = '/uploads/'. $uploadDir.'/thumbs/' . $filename;
+            $response['thumbUrl'] = '/uploads/'. $config['uploadDir'].'/thumbs/' . $filename;
             $response['size'] = filesize($finalPath);
+        }
+        else{
+            $response['thumbUrl'] = '/icon/'. $file->guessExtension();
         }
 
         return $this->json($response);
@@ -115,25 +115,27 @@ class GalerieUploadController extends AbstractController
 
 
     #[Route('/admin/upload/{data_type}/remove/{id}', name: 'admin_upload_remove', methods: ['DELETE'])]
-    public function delete(Foto $foto, EntityManagerInterface $em): JsonResponse
+    public function delete(string $data_type, int $id, EntityManagerInterface $em): JsonResponse
     {
         $baseDir = $this->getParameter('kernel.project_dir') . '/public/';
+        $fileEntity = $em->getRepository($this->mapFileEntity[$data_type] ?? null)->find($id);
 
         // originál
-        $filePath = $baseDir . $foto->getSoubor();
+        $filePath = $baseDir . $fileEntity->getSoubor();
 
         if (file_exists($filePath)) {
             unlink($filePath);
         }
 
-        // 🔥 thumbnail
-        $thumbPath = $baseDir . 'uploads/images/thumbs/' . basename($foto->getSoubor());
+        if ($data_type == 'image') {
+            // thumbnail
+            $thumbPath = $baseDir . 'uploads/' . $this->mapUploadDir[$data_type] . '/thumbs/' . basename($fileEntity->getSoubor());
 
-        if (file_exists($thumbPath)) {
-            unlink($thumbPath);
+            if (file_exists($thumbPath)) {
+                unlink($thumbPath);
+            }
         }
-
-        $em->remove($foto);
+        $em->remove($fileEntity);
         $em->flush();
 
         return $this->json(['success' => true]);
@@ -143,6 +145,7 @@ class GalerieUploadController extends AbstractController
     #[Route('/admin/upload/{data_type}/list/{entity}/{id}', name: 'admin_upload_list', methods: ['GET'])]
     public function list(
         string $entity,
+        string $data_type,
         int $id,
         EntityManagerInterface $em
     ): JsonResponse {
@@ -152,7 +155,7 @@ class GalerieUploadController extends AbstractController
             return $this->json(['error' => 'Unknown entity'], 400);
         }
 
-        $object = $em->getRepository($this->mapEntity[$entity])->find($id);
+        $object = $em->getRepository($this->mapEntity[$entity])?->find($id);
 
         if (!$object) {
             return $this->json([]);
@@ -161,38 +164,45 @@ class GalerieUploadController extends AbstractController
         if (!$object instanceof FotoInterface) {
             return $this->json(['error' => 'Entita nepodporuje fotky'], 400);
         }
-
-        $fotos = $object->getFotos();
+        $getter = $this->mapGetter[$data_type];
+        $files = $object->$getter();
 
         $projectDir = $this->getParameter('kernel.project_dir');
 
         $data = [];
 
-        foreach ($fotos as $foto) {
-            $path = $foto->getSoubor();
+        foreach ($files as $file) {
+            $path = $file->getSoubor();
             $fullPath = $projectDir . '/public/' . $path;
 
-            $data[] = [
-                'id' => $foto->getId(),
-                'name' => $foto->getNazev(),
+
+
+            $item = [
+                'id' => $file->getId(),
+                'name' => $file->getNazev(),
                 'url' => '/' . $path,
-                'thumbUrl' => '/uploads/images/thumbs/' . basename($path),
                 'size' => file_exists($fullPath) ? filesize($fullPath) : 0,
             ];
+
+            if ($data_type === 'image') {
+                $item['thumbUrl'] = '/uploads/'.$this->mapUploadDir[$data_type].'/thumbs/' . basename($path);
+            }
+
+            $data[] = $item;
         }
 
         return $this->json($data);
     }
 
     #[Route('/admin/upload/{data_type}/reorder', name: 'admin_upload_reorder', methods: ['POST'])]
-    public function reorder(Request $request, EntityManagerInterface $em): JsonResponse
+    public function reorder(Request $request, EntityManagerInterface $em, string $data_type): JsonResponse
     {
         $data = json_decode($request->getContent(), true);
 
         foreach ($data as $item) {
-            $foto = $em->getRepository(Foto::class)->find($item['id']);
-            if ($foto) {
-                $foto->setPosition($item['position']);
+            $fileEntity = $em->getRepository($this->mapFileEntity[$data_type] ?? null)?->find($item['id']);
+            if ($fileEntity) {
+                $fileEntity->setPosition($item['position']);
             }
         }
 
@@ -203,17 +213,21 @@ class GalerieUploadController extends AbstractController
 
     #[Route('/admin/upload/{data_type}/rename/{id}', name: 'admin_upload_rename', methods: ['POST'])]
     public function rename(
-        Foto $foto,
+        string $data_type,
         Request $request,
+        int $id,
         EntityManagerInterface $em
     ): JsonResponse {
+
         $data = json_decode($request->getContent(), true);
 
         if (!isset($data['nazev'])) {
             return $this->json(['error' => 'Missing name'], 400);
         }
-
-        $foto->setNazev($data['nazev']);
+        $fileEntity = $em->getRepository($this->mapFileEntity[$data_type] ?? null)?->find($id);
+        if ($fileEntity) {
+            $fileEntity->setNazev($data['nazev']);
+        }
         $em->flush();
 
         return $this->json(['success' => true]);
